Listening to requests on port 8000 (terminal) .....http://127.0.0.1:8000/api
{
id: 0,
productName: "Fresh Avocados",
image: "🥑",
from: "Spain",
nutrients: "Vitamin B, Vitamin K",
quantity: "4 🥑",
price: "6.50",
organic: true,
description: "A ripe avocado yields to gentle pressure when held in the palm of the hand and squeezed. The fruit is not sweet, but distinctly and subtly flavored, with smooth texture. The avocado is popular in vegetarian cuisine as a substitute for meats in sandwiches and salads because of its high fat content. Generally, avocado is served raw, though some cultivars, including the common 'Hass', can be cooked for a short time without becoming bitter. It is used as the base for the Mexican dip known as guacamole, as well as a spread on corn tortillas or toast, served with spices."
},
{
id: 1,
productName: "Goat and Sheep Cheese",
image: "🧀",
from: "Portugal",
nutrients: "Vitamin A, Calcium",
quantity: "250g",
price: "5.00",
organic: false,
description: "Creamy and distinct in flavor, goat cheese is a dairy product enjoyed around the world. Goat cheese comes in a wide variety of flavors and textures, from soft and spreadable fresh cheese to salty, crumbly aged cheese. Although it’s made using the same coagulation and separation process as cheese made from cow’s milk, goat cheese differs in nutrient content."
},
{
id: 2,
productName: "Apollo Broccoli",
image: "🥦",
from: "Portugal",
nutrients: "Vitamin C, Vitamin K",
quantity: "3 🥦",
price: "5.50",
organic: true,
description: "Broccoli is known to be a hearty and tasty vegetable which is rich in dozens of nutrients. It is said to pack the most nutritional punch of any vegetable. When we think about green vegetables to include in our diet, broccoli is one of the foremost veggies to come to our mind. Broccoli is a cruciferous vegetable and part of the cabbage family, which includes vegetables such as Brussel sprouts and kale. Although the tastes are different, broccoli and these other vegetables are from the same family."
},
{